+++ packages/react/src/ClapButton/templates/Heart/index.tsx
@@ -1,25 +1,24 @@
 /* @jsx jsx */
 
-import { jsx } from 'theme-ui';
+import { jsx, ThemeUIStyleObject } from 'theme-ui';
 import { ClapButtonTemplateComponentProps } from '../..';
-import { ClapEmpty } from './icons/ClapEmpty';
-import { ClapFull } from './icons/ClapFull';
+import { ClapHeart } from './icons/ClapHeart';
 import { keyframes } from '@emotion/core';
 import { FC, useCallback, useState, useRef } from 'react';
 import { style } from './style';
 
 const baloonFade = keyframes({
   '0%': {
-    opacity: '0.5',
-    transform: 'translateY(-2.6em)',
+    opacity: '0.2',
+    transform: 'translateY(-3em)',
   },
   '72%': {
     opacity: '1',
-    transform: 'translateY(-3.4em)',
+    transform: 'translateY(-3.6em)',
   },
   '100%': {
     opacity: '0',
-    transform: 'translateY(-5em)',
+    transform: 'translateY(-5.2em)',
   },
 });
 
@@ -46,7 +45,7 @@ const iconScale = keyframes({
     transform: 'scale(1.1)',
   },
   '100%': {
-    transform: 'scale(1)',
+    transform: 'scale(1.05)',
   },
 });
 
@@ -58,7 +57,7 @@ const BALOON = {
 const TRIANGLES = {
   count: 5,
   size: 0.2,
-  color: '#ff365e',
+  color: 'secondary',
   durationMs: 300,
   animation: triangleFade,
 };
@@ -66,7 +65,7 @@ const TRIANGLES = {
 const CIRCLES = {
   count: 5,
   size: 0.05,
-  color: '#4086ff',
+  color: 'primary',
   durationMs: 500,
   animation: triangleFade,
 };
@@ -76,7 +75,7 @@ const ICON = {
   animation: iconScale,
 };
 
-export const Medium: FC<ClapButtonTemplateComponentProps> = ({
+export const Heart: FC<ClapButtonTemplateComponentProps> = ({
   isLoading,
   userClaps,
   totalClaps,
@@ -87,58 +86,56 @@ export const Medium: FC<ClapButtonTemplateComponentProps> = ({
 
   const triangles = [];
   for (let i = 0; i < TRIANGLES.count; i++) {
+    const trianglesStyle: ThemeUIStyleObject = {
+      width: '0',
+      height: '0',
+      opacity: '0',
+      borderLeft: `${TRIANGLES.size}em solid transparent`,
+      borderRight: `${TRIANGLES.size}em solid transparent`,
+      borderTop: `${TRIANGLES.size}em solid ${TRIANGLES.color}`,
+      position: 'absolute',
+      animation: animationActive
+        ? `${triangleFade} ${TRIANGLES.durationMs}ms ease forwards`
+        : null,
+    };
+
+    const trianglesContainerStyle = {
+      transform: `rotate(${(360 / TRIANGLES.count) * i}deg)`,
+      transformOrigin: `${TRIANGLES.size}em 0`,
+      marginLeft: `-${TRIANGLES.size}em`,
+    };
+
     triangles.push(
-      <div
-        key={i}
-        sx={{
-          transform: `rotate(${(360 / TRIANGLES.count) * i}deg)`,
-          transformOrigin: `${TRIANGLES.size}em 0`,
-          marginLeft: `-${TRIANGLES.size}em`,
-        }}
-      >
-        <div
-          sx={{
-            width: '0',
-            height: '0',
-            opacity: '0',
-            borderLeft: `${TRIANGLES.size}em solid transparent`,
-            borderRight: `${TRIANGLES.size}em solid transparent`,
-            borderTop: `${TRIANGLES.size}em solid ${TRIANGLES.color}`,
-            position: 'absolute',
-            animation: animationActive
-              ? `${triangleFade} ${TRIANGLES.durationMs}ms ease forwards`
-              : null,
-          }}
-        />
+      <div key={i} sx={trianglesContainerStyle}>
+        <div sx={trianglesStyle} />
       </div>
     );
   }
 
   const circles = [];
   for (let i = 0; i < CIRCLES.count; i++) {
+    const circlesStyle: ThemeUIStyleObject = {
+      width: '0',
+      height: '0',
+      opacity: '0',
+      padding: `${CIRCLES.size}em`,
+      backgroundColor: `${CIRCLES.color}`,
+      borderRadius: '100%',
+      position: 'absolute',
+      animation: animationActive
+        ? `${triangleFade} ${CIRCLES.durationMs}ms ease forwards`
+        : null,
+    };
+
+    const circleContainerStyle = {
+      transform: `rotate(${(360 / CIRCLES.count) * i}deg)`,
+      transformOrigin: `${CIRCLES.size}em 0`,
+      marginLeft: `-${CIRCLES.size}em`,
+    };
+
     circles.push(
-      <div
-        key={i}
-        sx={{
-          transform: `rotate(${(360 / CIRCLES.count) * i}deg)`,
-          transformOrigin: `${CIRCLES.size}em 0`,
-          marginLeft: `-${CIRCLES.size}em`,
-        }}
-      >
-        <div
-          sx={{
-            width: '0',
-            height: '0',
-            opacity: '0',
-            padding: `${CIRCLES.size}em`,
-            backgroundColor: `${CIRCLES.color}`,
-            borderRadius: '100%',
-            position: 'absolute',
-            animation: animationActive
-              ? `${triangleFade} ${CIRCLES.durationMs}ms ease forwards`
-              : null,
-          }}
-        />
+      <div key={i} sx={circleContainerStyle}>
+        <div sx={circlesStyle} />
       </div>
     );
   }
@@ -148,28 +145,33 @@ export const Medium: FC<ClapButtonTemplateComponentProps> = ({
   const handleClick = useCallback(
     (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
       setAnimationActive(true);
-
       if (timeoutId.current) {
         clearTimeout(timeoutId.current);
       }
-
       timeoutId.current = setTimeout(
         () => setAnimationActive(false),
         BALOON.durationMs
       );
-
       handlePress(e);
     },
     [handlePress]
   );
 
-  const Icon = userClaps && userClaps > 0 ? ClapFull : ClapEmpty;
-
   const iconStyle = {
     ...style.icon,
+    fill: userClaps && userClaps > 0 ? 'highlight' : 'icon',
     animation: animationActive
       ? `${ICON.animation} ${ICON.durationMs}ms ease forwards`
       : null,
+  };
+
+  const baloonStyle = {
+    ...style.baloon,
+    ...{
+      animation: animationActive
+        ? `${BALOON.animation} ${BALOON.durationMs}ms ease forwards`
+        : null,
+    },
   };
 
   return (
@@ -178,14 +180,7 @@ export const Medium: FC<ClapButtonTemplateComponentProps> = ({
         <div sx={style.centeredContainer}>
           <div
             key={timeoutId.current && timeoutId.current.toString()}
-            sx={{
-              ...style.baloon,
-              ...{
-                animation: animationActive
-                  ? `${BALOON.animation} ${BALOON.durationMs}ms ease forwards`
-                  : null,
-              },
-            }}
+            sx={baloonStyle}
           >
             {userClaps}
           </div>
@@ -193,9 +188,10 @@ export const Medium: FC<ClapButtonTemplateComponentProps> = ({
         <button onClick={handleClick} sx={style.button} disabled={isLoading}>
           <div sx={style.centeredContainer}> {triangles} </div>
           <div sx={style.centeredContainer}> {circles} </div>
-          <Icon sx={iconStyle} />
+          <ClapHeart sx={iconStyle} />
         </button>
       </div>
+
       {isCounterVisible && <div sx={style.counter}>{totalClaps}</div>}
     </div>
   );