+++ packages/react/src/Client/index.ts
@@ -186,7 +186,7 @@ export class Client {
     }
   ) {
     if (typeof window === 'undefined') {
-      throw 'Client is supposed to be used client-side only!';
+      throw new Error('Client is supposed to be used client-side only!');
     }
 
     const defaultHeaders: Record<string, string> = {
@@ -218,7 +218,14 @@ export class Client {
     const response = await fetch(url, requestInit);
 
     if (response.status < 200 || response.status >= 300) {
-      throw new ApiError(url, requestInit, response);
+      const errorMessages = await response.json();
+
+      throw new ApiError(
+        url,
+        requestInit,
+        response.status,
+        errorMessages.errors
+      );
     }
 
     return response.json();