+++ packages/react/src/LikeButton/templates/Chevron/theme.ts
@@ -1,0 +1,13 @@
+const theme: Record<string, any> = {
+  buttons: {
+    inactive: {
+      backgroundColor: 'background',
+    },
+    active: {
+      transition: 'all 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms',
+      backgroundColor: 'primary',
+    },
+  },
+};
+
+export default theme;