+++ packages/react/src/LikeButton/templates/Simple/style.ts
@@ -1,4 +1,4 @@
-import { ThemeUIStyleObject } from 'theme-ui';
+import { ThemeUIStyleObject } from '@theme-ui/css';
 
 export const style: Record<string, ThemeUIStyleObject> = {
   root: {
@@ -11,13 +11,12 @@ export const style: Record<string, ThemeUIStyleObject> = {
     ml: '.5em',
     color: 'text',
     fontFamily: 'body',
-    fontWeight: 'body',
   },
   icon: {
     height: '1em',
     display: 'block',
     margin: '0 auto',
-    fill: 'text',
+    fill: 'icon',
   },
   ring: {
     opacity: '0',
@@ -35,8 +34,10 @@ export const style: Record<string, ThemeUIStyleObject> = {
   button: {
     border: 'none',
     outline: 'none',
+    boxShadow: '0 0 0',
+    transform: 'translate3d(0,0,0)',
     cursor: 'pointer',
-    background: 'transparent',
+    bg: 'background',
     appearance: 'none',
     position: 'relative',
     fontSize: 'inherit',
@@ -46,5 +47,9 @@ export const style: Record<string, ThemeUIStyleObject> = {
     height: '3em',
     padding: '0',
     lineHeight: '1em',
+    ':hover': {
+      transition: 'all 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms',
+      opacity: 0.75,
+    },
   },
 };