+++ packages/react/src/LikeButton/templates/Twitter/style.ts
@@ -1,4 +1,4 @@
-import { ThemeUIStyleObject } from 'theme-ui';
+import { ThemeUIStyleObject } from '@theme-ui/css';
 
 export const style: Record<string, ThemeUIStyleObject> = {
   root: {
@@ -8,7 +8,15 @@ export const style: Record<string, ThemeUIStyleObject> = {
     alignItems: 'center',
   },
   counter: {
-    ml: '.2em',
+    ml: '.4em',
+    fontFamily: 'body',
+    color: 'text',
+  },
+  icon: {
+    height: '1.7em',
+    display: 'block',
+    margin: '0 auto',
+    animationTimingFunction: 'cubic-bezier(0.280, 0.840, 0.420, 1)',
   },
   button: {
     border: 'none',
@@ -53,11 +61,5 @@ export const style: Record<string, ThemeUIStyleObject> = {
     top: '50%',
     left: '50%',
     transform: 'translate(-50%, -50%)',
-  },
-  icon: {
-    height: '2em',
-    display: 'block',
-    margin: '0 auto',
-    animationTimingFunction: 'cubic-bezier(0.280, 0.840, 0.420, 1)',
   },
 };