+++ packages/react/src/UpdownButton/templates/Chevron/icons/ChevronIcon.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+
+interface Props {
+  className?: string;
+}
+
+export const ChevronIcon = ({ className }: Props) => (
+  <svg
+    id="Chevron"
+    className={className}
+    version="1.1"
+    xmlns="http://www.w3.org/2000/svg"
+    x="0px"
+    y="0px"
+    viewBox="0 0 512 512"
+    xmlSpace="preserve"
+  >
+    <path
+      d="M505.183,123.179c-9.087-9.087-23.824-9.089-32.912,0.002l-216.266,216.27L39.729,123.179
+  	c-9.087-9.087-23.824-9.089-32.912,0.002c-9.089,9.089-9.089,23.824,0,32.912L239.55,388.82c4.364,4.364,10.283,6.816,16.455,6.816
+  	c6.172,0,12.092-2.453,16.455-6.817l232.721-232.727C514.272,147.004,514.272,132.268,505.183,123.179z"
+    />
+  </svg>
+);