+++ packages/react/stories/UpdownButton.stories.tsx
@@ -1,52 +1,97 @@
 import * as React from 'react';
-import { Provider, UpdownButton } from '../src';
+import { UpdownButton } from '../src';
+import {
+  Provider,
+  ThemedProviderContrast,
+  ThemedProvider,
+} from './utils/testProvider';
 
 export default {
   title: 'UpdownButton',
 };
 
 export const Simple = () => (
-  <Provider
-    apiKey="xxx"
-    baseUrl="http://localhost:3000"
-    disableSessionId
-    theme={{
-      colors: {
-        primary: '#BFFCED',
-        secondary: '#FBDEFB',
-        text: 'grey',
-        background: '#eed',
-      },
-    }}
-  >
-    <>
-      <UpdownButton id="simple-example-sm" namespace="stories" />
-      <div style={{ fontSize: '20px' }}>
+  <>
+    <Provider>
+      <div
+        style={{
+          fontSize: '30px',
+          marginBottom: '40px',
+          display: 'inline',
+          fontFamily: 'Papyrus',
+        }}
+      >
+        <UpdownButton id="simple-example-sm" namespace="stories" />
+      </div>
+    </Provider>
+
+    <ThemedProvider>
+      <div style={{ marginBottom: '40px', fontSize: '40px' }}>
         <UpdownButton id="simple-example-md" namespace="stories" />
       </div>
-      <div style={{ fontSize: '30px' }}>
+    </ThemedProvider>
+
+    <ThemedProviderContrast>
+      <div style={{ marginBottom: '40px', fontSize: '60px' }}>
         <UpdownButton id="simple-example-bg" namespace="stories" />
       </div>
-    </>
-  </Provider>
+    </ThemedProviderContrast>
+  </>
+);
+
+export const Chevron = () => (
+  <>
+    <Provider>
+      <div style={{ marginBottom: '40px', fontSize: '30px' }}>
+        <UpdownButton
+          id="simple-example-sm"
+          namespace="stories"
+          component={UpdownButton.templates.Chevron}
+        />
+      </div>
+    </Provider>
+
+    <ThemedProvider>
+      <div style={{ marginBottom: '40px', fontSize: '40px' }}>
+        <UpdownButton
+          id="simple-example-md"
+          namespace="stories"
+          component={UpdownButton.templates.Chevron}
+        />
+      </div>
+    </ThemedProvider>
+
+    <ThemedProviderContrast>
+      <div style={{ marginBottom: '40px', fontSize: '60px' }}>
+        <UpdownButton
+          id="simple-example-bg"
+          namespace="stories"
+          component={UpdownButton.templates.Chevron}
+        />
+      </div>
+    </ThemedProviderContrast>
+  </>
 );
 
 export const Reddit = () => (
-  <Provider apiKey="xxx" baseUrl="http://localhost:3000">
+  <Provider>
     <>
-      <UpdownButton
-        id="reddit-example-sm"
-        namespace="stories"
-        component={UpdownButton.templates.Reddit}
-      />
-      <div style={{ fontSize: '20px' }}>
+      <div style={{ marginBottom: '40px', fontSize: '30px' }}>
+        <UpdownButton
+          id="reddit-example-sm"
+          namespace="stories"
+          component={UpdownButton.templates.Reddit}
+        />
+      </div>
+
+      <div style={{ marginBottom: '40px', fontSize: '20px' }}>
         <UpdownButton
           id="reddit-example-md"
           namespace="stories"
           component={UpdownButton.templates.Reddit}
         />
       </div>
-      <div style={{ fontSize: '30px' }}>
+      <div style={{ marginBottom: '40px', fontSize: '30px' }}>
         <UpdownButton
           id="reddit-example-bg"
           namespace="stories"
@@ -58,7 +103,7 @@ export const Reddit = () => (
 );
 
 export const Custom = () => (
-  <Provider apiKey="xxx" baseUrl="http://localhost:3000">
+  <Provider>
     <UpdownButton id="custom" namespace="stories">
       {({
         handlePressUp,